Combines JSONSubscriptTests and JSONSubscriptingTests
