+++ Tests/JSONSubscriptingTests.swift
@@ -11,23 +11,49 @@ import Freddy
 
 class JSONSubscriptingTests: XCTestCase {
     
-    var json: JSON!
+    private var residentJSON: JSON!
+    private var json: JSON!
+    private var noWhiteSpaceData: NSData!
+    
+    func parser() -> JSONParserType.Type {
+        return JSONParser.self
+    }
     
     override func setUp() {
         super.setUp()
         
-        json = JSON.Dictionary([
+        residentJSON = JSON.Dictionary([
             "residents": [
                 ["name": "Matt", "age": 33, "hasPet": false, "rent": .Null],
                 ["name": "Drew", "hasPet": true, "rent": 1234.5],
                 ["name": "Pat", "age": 28, "hasPet": .Null]
             ]
             ])
+        
+        let testBundle = NSBundle(forClass: JSONSubscriptingTests.self)
+        guard let data = testBundle.URLForResource("sample", withExtension: "JSON").flatMap(NSData.init) else {
+            XCTFail("Could not read sample data from test bundle")
+            return
+        }
+        
+        do {
+            self.json = try JSON(data: data, usingParser: parser())
+        } catch {
+            XCTFail("Could not parse sample JSON: \(error)")
+            return
+        }
+        
+        guard let noWhiteSpaceData = testBundle.URLForResource("sampleNoWhiteSpace", withExtension: "JSON").flatMap(NSData.init) else {
+            XCTFail("Could not read sample data (no whitespace) from test bundle")
+            return
+        }
+        
+        self.noWhiteSpaceData = noWhiteSpaceData
     }
     
     func testThatArrayOfProducesResidents() {
         do {
-            let residents = try json.arrayOf("residents", type: Resident.self)
+            let residents = try residentJSON.arrayOf("residents", type: Resident.self)
             let residentsNames = residents.map { $0.name }
             XCTAssertEqual(residentsNames.count, 3, "There should be 3 residents.")
         } catch {
@@ -37,7 +63,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testNullOptionsDecodes() {
         do {
-            let firstResident = try json.decode("residents", 0, alongPath: [.NullBecomesNil, .MissingKeyBecomesNil], type: Resident.self)
+            let firstResident = try residentJSON.decode("residents", 0, alongPath: [.NullBecomesNil, .MissingKeyBecomesNil], type: Resident.self)
             XCTAssertNotNil(firstResident)
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -46,7 +72,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testNullOptionsProducesOptionalForNotFound() {
         do {
-            let residents = try json.arrayOf("residents", type: Resident.self)
+            let residents = try residentJSON.arrayOf("residents", type: Resident.self)
             XCTAssertNil(residents[1].age, "Drew's `age` should be nil.")
         } catch {
             XCTFail("There should be no error.")
@@ -55,7 +81,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testNullOptionsProducesOptionalForNullOrNotFound() {
         do {
-            let residents = try json.arrayOf("residents", type: Resident.self)
+            let residents = try residentJSON.arrayOf("residents", type: Resident.self)
             XCTAssertNil(residents.first?.rent, "Matt should have nil `rent`.")
             XCTAssertEqual(residents[1].rent!, 1234.5, "Drew's `rent` should equal 1234.5.")
             XCTAssertNil(residents.last?.rent, "Pat should have nil `rent`.")
@@ -66,7 +92,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testNullOptionsIndexOutOfBoundsProducesOptional() {
         do {
-            let residentOutOfBounds = try json.decode("residents", 4, alongPath: .MissingKeyBecomesNil, type: Resident.self)
+            let residentOutOfBounds = try residentJSON.decode("residents", 4, alongPath: .MissingKeyBecomesNil, type: Resident.self)
             XCTAssertNil(residentOutOfBounds)
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -111,7 +137,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testJSONDictionaryUnexpectedSubscript() {
         do {
-            _ = try json.decode(1, type: Resident.self)
+            _ = try residentJSON.decode(1, type: Resident.self)
             XCTFail("Should throw error.")
         } catch JSON.Error.UnexpectedSubscript(type: let theType) {
             XCTAssertTrue(theType is Int.Type)
@@ -121,18 +147,18 @@ class JSONSubscriptingTests: XCTestCase {
     }
     
     func testJSONIndexSubscript() {
-        let residents = json["residents"]
+        let residents = residentJSON["residents"]
         XCTAssertNotNil(residents)
     }
     
     func testJSONKeySubscript() {
-        let matt = json["residents"]?[0]
+        let matt = residentJSON["residents"]?[0]
         XCTAssertNotNil(matt)
     }
     
     func testDecodeOr() {
         do {
-            let outOfBounds = try json.decode("residents", 4, or: Resident(name: "NA", age: 30, hasPet: false, rent: 0))
+            let outOfBounds = try residentJSON.decode("residents", 4, or: Resident(name: "NA", age: 30, hasPet: false, rent: 0))
             XCTAssertTrue(outOfBounds.name == "NA")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -141,7 +167,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testDoubleOr() {
         do {
-            let rent = try json.double("residents", 2, "rent", or: 0)
+            let rent = try residentJSON.double("residents", 2, "rent", or: 0)
             XCTAssertTrue(rent == 0, "Rent should be free.")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -150,7 +176,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testStringOr() {
         do {
-            let nickname = try json.string("residents", 0, "nickname", or: "DubbaDubs")
+            let nickname = try residentJSON.string("residents", 0, "nickname", or: "DubbaDubs")
             XCTAssertTrue(nickname == "DubbaDubs")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -159,7 +185,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testIntOr() {
         do {
-            let age = try json.int("residents", 1, "age", or: 21)
+            let age = try residentJSON.int("residents", 1, "age", or: 21)
             XCTAssertTrue(age == 21, "Forever young!")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -168,7 +194,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testBoolOr() {
         do {
-            let hasSpouse = try json.bool("residents", 1, "hasSpouse", or: false)
+            let hasSpouse = try residentJSON.bool("residents", 1, "hasSpouse", or: false)
             XCTAssertFalse(hasSpouse, "No spouse")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -189,7 +215,7 @@ class JSONSubscriptingTests: XCTestCase {
     func testArrayOfOr() {
         do {
             let matt = Resident(name: "Matt", age: 32, hasPet: false, rent: 500.00)
-            let residentsOr = try json.arrayOf("residnts", or: [matt])
+            let residentsOr = try residentJSON.arrayOf("residnts", or: [matt])
             XCTAssertEqual(residentsOr.first!, matt, "`residents` should not be nil")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -199,7 +225,7 @@ class JSONSubscriptingTests: XCTestCase {
     func testDictionaryOr() {
         do {
             let jsonDict: [String: JSON] = ["name": "Matt", "age": 33, "hasPet": false, "rent": .Null]
-            let mattOr = try json.dictionary("residents", 4, or: jsonDict)
+            let mattOr = try residentJSON.dictionary("residents", 4, or: jsonDict)
             XCTAssertEqual(jsonDict, mattOr, "`jsonDict` should equal `mattOr`")
         } catch {
             XCTFail("There should be no error: \(error).")
@@ -208,7 +234,7 @@ class JSONSubscriptingTests: XCTestCase {
     
     func testThatUnexpectedSubscriptIsThrown() {
         do {
-            _ = try json.decode("residents", 1, "name", "initial", type: Resident.self)
+            _ = try residentJSON.decode("residents", 1, "name", "initial", type: Resident.self)
         } catch JSON.Error.UnexpectedSubscript(let type) {
             XCTAssert(type == Swift.String, "The dictionary at index 1 should not be subscriptable by: \(type).")
         } catch {
@@ -233,7 +259,7 @@ class JSONSubscriptingTests: XCTestCase {
         }
     }
     
-    func testNullOptionsStillFailIfKeyIsMissing() {
+    func testParsingOptionsStillFailIfKeyIsMissing() {
         do {
             let json = JSON.Dictionary([
                 "name": "Drew",
@@ -246,6 +272,83 @@ class JSONSubscriptingTests: XCTestCase {
         } catch {
             XCTFail("Unexpected error: \(error)")
         }
+    }
+    
+    func testThatMapCanCreateArrayOfPeople() {
+        let peopleJSON = try! json.array("people")
+        let people = try! peopleJSON.map(Person.init)
+        for person in people {
+            XCTAssertNotEqual(person.name, "", "There should be a name.")
+        }
+    }
+    
+    func testThatSubscriptingJSONWorksForTopLevelObject() {
+        let success = try? json.bool("success")
+        XCTAssertEqual(success, true, "There should be `success`.")
+    }
+    
+    func testThatPathSubscriptingPerformsNesting() {
+        for z in try! json.array("states", "Georgia") {
+            XCTAssertNotNil(try? z.int(), "The `Int` should not be `nil`.")
+        }
+    }
+    
+    func testJSONSubscriptWithInt() {
+        let mattmatt = try? json.string("people", 0, "name")
+        XCTAssertEqual(mattmatt, "Matt Mathias", "`matt` should hold string `Matt Mathias`")
+    }
+    
+    func testJSONErrorKeyNotFound() {
+        do {
+            _ = try json.array("peopl")
+        } catch JSON.Error.KeyNotFound(let key) {
+            XCTAssert(key == "peopl", "The error should be due to the key not being found.")
+        } catch {
+            XCTFail("The error should be due to the key not being found, but was: \(error).")
+        }
+    }
+    
+    func testJSONErrorIndexOutOfBounds() {
+        do {
+            _ = try json.dictionary("people", 4)
+        } catch JSON.Error.IndexOutOfBounds(let index) {
+            XCTAssert(index == 4, "The error should be due to the index being out of bounds.")
+        } catch {
+            XCTFail("The error should be due to the index being out of bounds, but was: \(error).")
+        }
+    }
+    
+    func testJSONErrorTypeNotConvertible() {
+        do {
+            _ = try json.int("people", 0, "name")
+        } catch let JSON.Error.ValueNotConvertible(value, to) {
+            XCTAssert(to == Swift.Int, "The error should be due the value not being an `Int` case, but was \(to).")
+            XCTAssert(value == "Matt Mathias", "The error should be due the value being the String 'Matt Mathias', but was \(value).")
+        } catch {
+            XCTFail("The error should be due to `name` not being convertible to `int`, but was: \(error).")
+        }
+    }
+    
+    func testJSONErrorUnexpectedSubscript() {
+        do {
+            _ = try json.string("people", "name")
+        } catch JSON.Error.UnexpectedSubscript(let type) {
+            XCTAssert(type == Swift.String, "The error should be due the value not being subscriptable with string `String` case, but was \(type).")
+        } catch {
+            XCTFail("The error should be due to the `people` `Array` not being subscriptable with `String`s, but was: \(error).")
+        }
+    }
+    
+    func testThatOptionalSubscriptingIntoNullSucceeds() {
+        let earlyNull = [ "foo": nil ] as JSON
+        let string = try! earlyNull.string("foo", "bar", "baz", alongPath: .NullBecomesNil)
+        XCTAssertNil(string)
+    }
+    
+    func testThatOptionalSubscriptingKeyNotFoundSucceeds() {
+        let keyNotFound = [ "foo": 2 ] as JSON
+        let string = try! keyNotFound.string("bar", alongPath: .MissingKeyBecomesNil)
+        XCTAssertNil(string)
     }
     
 }
@@ -273,4 +376,39 @@ private func ==(lhs: Resident, rhs: Resident) -> Bool {
     (lhs.age == rhs.age) &&
     (lhs.hasPet == rhs.hasPet) &&
     (lhs.rent == rhs.rent)
+}
+
+class JSONSubscriptWithNSJSONTests: JSONSubscriptingTests {
+    
+    override func parser() -> JSONParserType.Type {
+        return NSJSONSerialization.self
+    }
+    
+}
+
+// Just for syntax validation, not for execution or being counted for coverage.
+private func testUsage() {
+    let j = JSON.Null
+    
+    _ = try? j.int()
+    _ = try? j.int(alongPath: .MissingKeyBecomesNil)
+    _ = try? j.int(alongPath: .NullBecomesNil)
+    _ = try? j.int(or: 42)
+    
+    _ = try? j.int("key")
+    _ = try? j.int("key", alongPath: .MissingKeyBecomesNil)
+    _ = try? j.int("key", alongPath: .NullBecomesNil)
+    _ = try? j.int("key", or: 42)
+    
+    _ = try? j.int(1)
+    _ = try? j.int(2, alongPath: .MissingKeyBecomesNil)
+    _ = try? j.int(3, alongPath: .NullBecomesNil)
+    _ = try? j.int(4, or: 42)
+    
+    let stringConst = "key"
+    
+    _ = try? j.int(stringConst, 1)
+    _ = try? j.int(stringConst, 2, alongPath: .MissingKeyBecomesNil)
+    _ = try? j.int(stringConst, 3, alongPath: .NullBecomesNil)
+    _ = try? j.int(stringConst, 4, or: 42)
 }
